Moved container clases to archive_types.py
